Avoid analyzing API example code twice, clean-up (#103548)
