+++ packages/flutter/lib/src/widgets/page_view.dart
@@ -108,7 +108,6 @@ import 'viewport.dart';
 ///     );
 ///   }
 /// }
-///
 /// ```
 /// {@end-tool}
 class PageController extends ScrollController {
